SS-8135: separated packages into headless and hydra viewer API
