+++ components/shapediver/viewport/ViewportComponent.tsx
@@ -3,7 +3,7 @@ import { useViewport } from "../../../hooks/shapediver/viewer/useViewport";
 import { ViewportCreateDto } from "../../../types/store/shapediverStoreViewport";
 import classes from "./ViewportComponent.module.css";
 import { MantineThemeComponent, useComputedColorScheme, useProps } from "@mantine/core";
-import { BUSY_MODE_DISPLAY, SPINNER_POSITIONING } from "@shapediver/viewer";
+import { BUSY_MODE_DISPLAY, SPINNER_POSITIONING } from "@shapediver/viewer.viewport";
 import AlertPage from "../../../pages/misc/AlertPage";
