+++ hooks/shapediver/viewer/interaction/selection/useHoverManagerEvents.ts
@@ -1,4 +1,4 @@
-import { addListener, EVENTTYPE_INTERACTION, IEvent, removeListener } from "@shapediver/viewer";
+import { addListener, EVENTTYPE_INTERACTION, IEvent, removeListener } from "@shapediver/viewer.session";
 import { InteractionEventResponseMapping, matchNodesWithPatterns, OutputNodeNameFilterPatterns } from "@shapediver/viewer.features.interaction";
 import { useState, useEffect } from "react";
